Adding in-cradle masking to volume-up (B-04547)
Activated on Rev-B phones only. When a RevB phone is cradled,
the hardware mistakes this for the case designed to boot unpowered
phones and electronically injects a 2-second volume-up press. This
unfortunately preceeds the cradle-detect GPIO line by somewhere between
85 and 100 ms. So we hold off on volume-up for 100 ms before polling
both the volume-up and the in-cradle gpio pins.
  If you press volume-up while the phone is even resting in the cradle,
it is ignored. If you can press volume-up fast enough (< 0.1 seconds),
it is also ignored. Sigh.

Reviewed by Hardik.